Feat:api domain별 폴더정리
