+++ apps/sfaclog/src/utils/api/lecture.ts
@@ -1,0 +1,1 @@
+import pb from "../pocketbase"
