+++ apps/sfaclog/src/utils/api/log.ts
@@ -1,0 +1,1 @@
+import pb from "../pocketbase"
